疯狂修改

+++ MVCiHealth/Models/iHealthModel.edmx.sql
@@ -75,7 +75,7 @@ CREATE TABLE [dbo].[DOCTOR] (
     [SECTION_ID] int  NULL,
     [DISEASE_ID] int  NULL,
     [INTRODUCTION] nvarchar(100)  NULL,
-	[LEVEL] int NULL,
+	[LEVEL] float NULL,
     [INSDATE] datetime  NULL
 );
 GO
@@ -85,7 +85,7 @@ CREATE TABLE [dbo].[DOCTOR_EVALUATION] (
     [EVALUATION_ID] int  NOT NULL,
     [PATIENT_ID] int  NOT NULL,
     [DOCTOR_ID] int  NOT NULL,
-    [RATE] int  NULL,
+    [RATE] float  NULL,
     [DETAIL] nvarchar(100)  NULL,
     [AGREETIMES] int  NOT NULL,
     [INSDATE] datetime  NOT NULL
